Refactor AppWrapper
